backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
pydantic-settings==2.1.0
httpx==0.26.0
requests==2.31.0
apscheduler==3.10.4
playwright==1.41.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
websockets==12.0

# Development dependencies
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
